backend/requirements.txt
# Claude Agent SDK - SDK oficial para criar agentes com Claude
claude-agent-sdk

# FastAPI e servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# MCP Server
fastmcp>=0.1.0

# Seguranca (Semana 1)
slowapi>=0.1.9            # Rate limiting
